requirements.txt
# Dependencies for Basic-3D-Warp-Simulation/example.py
# Install with: python -m pip install -r requirements.txt

numpy
pygame
PyOpenGL
PyOpenGL_accelerate  # optional: can speed up PyOpenGL

# NVIDIA Warp (used as `import warp as wp` in the script):
# Warp requires a CUDA-enabled build to use device="cuda" as in the example.
# The exact wheel or install method depends on your CUDA version and OS.
# Try:
#   python -m pip install warp
# If that fails, follow NVIDIA Warp's installation instructions for your CUDA version
# (you may need to install from a vendor index or build from source).
# If you don't have a CUDA GPU or cannot install the CUDA build, change the script
# to use device="cpu" (or remove device argument) and consult Warp docs for CPU support.

# Notes:
# - Make sure you have up-to-date pip/setuptools/wheel before installing binary wheels:
#     python -m pip install --upgrade pip setuptools wheel
# - If you run into permission issues on Windows, run PowerShell as Administrator or
#   use the --user flag with pip.
